kas: add runqemu wrapper
Add a wrapper script to run runqemu using a local kas-built build tree.

This is necessary because runqemu won't work correctly within the kas
build container (because runqemu wants to set up networking and doesn't
have enough stuff inside the container to do this correctly), and it
won't work correctly outside the container either (because on many
codepaths it wants to call `bitbake -e`, and we don't have bitbake in
the host environment).

Most of what runqemu needs bitbake for is to find a *.qemuboot.conf;
this script handles that (using kas project file information) and calls
runqemu to handle the rest.

Signed-off-by: Brenda Streiff <brenda.streiff@emerson.com>