kas: patches: add patch to fix TARGET_GOARCH expansion error
A recent oe-core merge from master to scarthgap introduced a breaking
change to parsing for every recipe that uses go that is parsed with a
microblazeel machine target (such as with versal), regardless of whether
that recipe is built for that target or not.

The error manifests as a "KeyError: 'Cannot map architecture microblazeel'"
buried inside of an ExpansionError.

https://lists.openembedded.org/g/openembedded-core/topic/116430143#msg226817

Revert the offending commit for now. I am expecting that a proper fix for
this will show up soon either in upstream oe-core or in meta-xilinx.

Signed-off-by: Brenda Streiff <brenda.streiff@emerson.com>